apps/api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.35
pandas>=2.2.3
numpy>=1.26.4
scikit-learn>=1.5.2
xgboost>=2.1.1
joblib>=1.4.2
scipy>=1.14.0
httpx>=0.27.2
pydantic-settings>=2.5.2
google-genai>=1.0.0
python-dotenv>=1.0.1
